requirements.txt
selenium
pytz
psutil
pyvirtualdisplay
mysql-connector
free-proxy
pandas
flask
flask_basicauth
flask_session
gunicorn
geoip2
geocoder
bs4
webdriver-manager
free-proxy
